Create WalletLogs component
